requirements.txt
httpx==0.27.2
playwright>=1.45
pillow>=10.0
